Go back to 1 byte for response data.
At least the code got some cleaning up. :)
#71

+++ test/com/esotericsoftware/kryonet/rmi/RmiTest.java
@@ -126,7 +126,7 @@ public class RmiTest extends KryoNetTestCase {
 				assertEquals(other, remoteObject.waitForLastResponse());
 
 				assertEquals(0f, test.other());
-				short responseID = remoteObject.getLastResponseID();
+				byte responseID = remoteObject.getLastResponseID();
 				assertEquals(other, remoteObject.waitForResponse(responseID));
 
 				// Non-blocking call that errors out
